Trivial comments

+++ engine/src/flutter/impeller/entity/shaders/rrect_blur.frag
@@ -31,7 +31,7 @@ frag_info;
 
 in vec2 v_position;
 
-out f16vec4 frag_color;
+out vec4 frag_color;
 
 void main() {
   vec2 centered = abs(v_position - frag_info.center);
